Converted volume to an integer

+++ src/main/java/com/korbyte/alphavantage/core/daily/DailyPositionData.java
@@ -12,5 +12,5 @@ import lombok.ToString;
 public class DailyPositionData extends BasePositionData {
 
   @JsonProperty(required = true, value = "5. volume")
-  private Double volume;
+  private Integer volume;
 }
